+++ xpaw/__init__.py
@@ -10,14 +10,14 @@ from .downloader import Downloader
 from .spider import Spider, every
 from .selector import Selector
 from .item import Item, Field
-from .run import run_spider, run_crawler
+from .run import run_spider, run_crawler, make_requests
 
-__all__ = ('FormData',
+__all__ = ['FormData',
            'HttpRequest', 'HttpResponse',
            'Downloader',
            'Spider', 'every',
            'Selector',
            'Item', 'Field',
-           'run_spider', 'run_crawler')
+           'run_spider', 'run_crawler', 'make_requests']
 
 del _patch